requirements.txt
openai
python-dotenv
tenacity
regex
colorlog
numpy
scikit-learn
sentence-transformers
openbox
